snn v1.1
Remove the drawing and data structure from the main code file. Modify the calculation method, no longer use list multiplication, but calculate by recording the signal length. The set and reset codes are added to the actual test results and errors, and the errors of the second repeatability test results are not added yet